+++ src/app/filter/filter-overview/filter-overview.component.ts
@@ -5,29 +5,22 @@
  */
 
 import { Component, ErrorHandler, OnInit } from "@angular/core";
-import { TableDataSource } from "@abraxas/base-components";
-import { FilterColumn } from "./filter-column";
 import { FilterService } from "../../services/filter.service";
 import { FilterDefinition } from "../../models/filter/filterDefinition";
 import { ActivatedRoute, Router } from "@angular/router";
 import { RoleService } from "src/app/services/role.service";
+import { AuthorizationService, Tenant } from "@abraxas/base-components";
 
 @Component({
   selector: 'app-filter-overview',
   templateUrl: './filter-overview.component.html',
-  styleUrls: ['./filter-overview.component.scss']
+  styleUrls: ['./filter-overview.component.scss'],
 })
 export class FilterOverviewComponent implements OnInit {
-
   public isManager: boolean = false;
   public loaded: boolean = false;
-  public columns = FilterColumn;
-  public columnsToDisplay: string[] = [
-    FilterColumn.NAME,
-    FilterColumn.DESCRIPTION,
-    FilterColumn.LATEST_VERSION,
-  ];
-  public filters = new TableDataSource<FilterDefinition>();
+  public myFilters: FilterDefinition[] = [];
+  public aggregateFilters: FilterDefinition[] = [];
 
   constructor(
     private readonly filterService: FilterService,
@@ -35,23 +28,24 @@ export class FilterOverviewComponent implements OnInit {
     private readonly router: Router,
     private readonly route: ActivatedRoute,
     private readonly errorHandler: ErrorHandler,
-  ) {
-  }
+    private readonly authorization: AuthorizationService
+  ) {}
 
   public async ngOnInit(): Promise<void> {
+    const tenant = await this.authorization.getActiveTenant();
     this.isManager = await this.roleService.isManager();
-    this.loadData();
+    this.loadData(tenant);
   }
 
-  private loadData(): void {
-    this.filterService.getAll().then(filters => {
-      this.filters.data = filters;
-      this.loaded = true;
-    }).catch(e => this.errorHandler.handleError(e));
-  }
-
-  public async openDetailView(filter: FilterDefinition): Promise<void> {
-    await this.router.navigate([filter.id], {relativeTo: this.route});
+  private loadData(tenant: Tenant): void {
+    this.filterService
+      .getAll()
+      .then((filters) => {
+        this.myFilters = filters.filter((f) => (f.tenantId === tenant.id));
+        this.aggregateFilters = filters.filter((f) => (f.tenantId !== tenant.id));
+        this.loaded = true;
+      })
+      .catch((e) => this.errorHandler.handleError(e));
   }
 
   public async openCreateView(): Promise<void> {