+++ src/app/models/registration/registrationStatisticResponse.model.ts
@@ -1,0 +1,12 @@
+/**
+ * (c) Copyright by Abraxas Informatik AG
+ *
+ * For license information see LICENSE file.
+ */
+
+import { RegistrationStatistic } from './registrationStatistic';
+
+export interface RegistrationStatisticResponseModel {
+  registrationStatistics: RegistrationStatistic[];
+  totalCount: number;
+}