+++ src/app/registration/registration-statistics/registration-statistics.component.ts
@@ -1,0 +1,14 @@
+/**
+ * (c) Copyright by Abraxas Informatik AG
+ *
+ * For license information see LICENSE file.
+ */
+
+import { Component } from '@angular/core';
+
+@Component({
+  selector: 'app-registration-statistics',
+  templateUrl: './registration-statistics.component.html',
+  styleUrls: ['./registration-statistics.component.scss'],
+})
+export class RegistrationStatisticsComponent {}