+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-# ✨ Changelog (`v1.77.15`)
+# ✨ Changelog (`v1.79.1`)
 
 All notable changes to this project will be documented in this file.
 
@@ -8,11 +8,41 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## Version Info
 
 ```text
-This version -------- v1.77.15
+This version -------- v1.79.1
 Previous version ---- v1.77.15
 Initial version ----- v1.77.15
-Total commits ------- 1
+Total commits ------- 5
 ```
+
+## [v1.79.1] - 2024-04-03
+
+### 🔄 Changed
+
+- fix checkbox initialization in filter
+
+## [v1.79.0] - 2024-03-26
+
+### 🆕 Added
+
+- add registration statistics
+
+## [v1.78.2] - 2024-03-15
+
+### 🔄 Changed
+
+- extend input validation rules
+
+## [v1.78.1] - 2024-02-14
+
+### 🔄 Changed
+
+- use absolut path for config.js script
+
+## [v1.78.0] - 2024-02-06
+
+### 🆕 Added
+
+- add filter tenant name
 
 ## [v1.77.15] - 2024-01-31